+++ docs/src/dealer/dealer.ts
@@ -10,6 +10,10 @@ export class Dealer extends Role {
         super(Dealer.name.toLocaleLowerCase());
     }
 
+    private get hiddenCard(): HTMLImageElement {
+        return document.querySelectorAll("[id$=hidden]")[0] as HTMLImageElement;
+    }
+
     protected writeCard(card: Card): void {
         super.writeCard(card, this.isInitialHand);
     }
@@ -43,16 +47,26 @@ export class Dealer extends Role {
         return this.hand.length <= 2;
     }
 
-    // TODO
-    checkBlackjack(): boolean {
+    async checkBlackjack(): Promise<boolean> {
         const card = this.hand[0];
 
-        if((Rank.ACE === card.rank) || (card.dictionary.get(card.rank) === 10)) {
-            // checking for blackjack...
+        if ((Rank.ACE === card.rank) || (card.dictionary.get(card.rank) === 10)) {
+            const checkMessage = document.createElement('span');
 
-            return this.blackjack;
+            checkMessage.textContent = 'The dealer is checking if he has blackjack...'
+            checkMessage.className = 'dealer-check';
+            this.hiddenCard.style.opacity = '0.5';
+
+            this.role.append(checkMessage);
+
+            return new Promise(resolve => setTimeout(resolve, 5000))
+                .then(() => checkMessage.textContent = this.blackjack ? 'The dealer has blackjack' : 'The dealer doesn\'t have blackjack')
+                .then(() => new Promise(resolve => setTimeout(resolve, 5000)))
+                .then(() => this.role.removeChild(checkMessage))
+                .then(() => this.hiddenCard.style.opacity = '1')
+                .then(() => true);
         }
 
-        return false;
+        return Promise.resolve(false);
     }
 }