+++ CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [UNRELEASED]
+## [1.3.0]
+
+### Added
+- Check if the dealer has blackjack.
 
 ### Modified
 - Hidden dealer cards at round start.
 
-## [1.2.1] - 15/08/2025
+## [1.2.0] - 15/08/2025
 
 ### Added
 - Betting chips.